fix broken link on styling in react
Visiting the most recent copy of the original link on the wayback machine, it's [this](http://web.archive.org/web/20190111132559/http://andrewhfarmer.com/how-to-style-react) which redirects to [here](http://web.archive.org/web/20180429195243/https://www.javascriptstuff.com/how-to-style-react/). The website it redirects to is still online and doesn't need the wayback maching to view, this commit uses the new link.